Temporarily force "serif" to be resolved as "Times"
Since 2cc5442b0263a449f4f5eb0cbf99e419ee0d0d38, in Qt5, the default
serif family picked by fontconfig in DumpRenderTree has changed.
If we want to be able to keep the tests results in line with what we
get for Qt4, we need this little hack to force the same family to be
resolved.