Add one more aliasing rule to not mess with the Qt4 baseline.
The approach consisting in hard coding font families in
QFont::defaultFont resulted in interesting results.
While the default font was always intended to be a serif font,
it returned "Times" which somehow resolved to "Nimbus Sans L".
Since QFont::defaultFont now returns "Nimbus Roman No9 L"
early on, we need this additional alias so that it later on gets
the same test font.